Add stack-friendly functions
Some functions declare relatively large data structures (such as
SHA3 contexts and bundle contexts) as local variables; this could
put excessive pressure on the stack in memory-constrained devices.
This commit adds variations to these functions that instead of
declaring the data structures as local variables take as additional
arguments pointers to caller-allocated structures; this allows
callers to allocate the structures wherever is best for the
platform being used.